requirements.txt
dnspython~=2.7.0
requests~=2.32.3
Kivy~=2.3.1
Buildozer
PyInstaller